btrfs-progs: Properly cast to avoid compiler warnings, fixes FTBFS on alpha and ia64
Bug-Debian: http://bugs.debian.org/539433
Bug-Debian: http://bugs.debian.org/583768
Authors:
 Luca Bruno <lucab@debian.org>
 Alexander Kurtz <kurtz.alex@googlemail.com>
 Daniel Baumann <daniel.baumann@progress-technologies.net>

+++ btrfs-convert.c
@@ -2441,7 +2441,7 @@ static int do_rollback(const char *devname)
 	ext2_root = btrfs_read_fs_root(root->fs_info, &key);
 	if (!ext2_root || IS_ERR(ext2_root)) {
 		fprintf(stderr, "unable to open subvol %llu\n",
-			key.objectid);
+			(unsigned long long)key.objectid);
 		goto fail;
 	}
